[refactor] 패키징 정리(#32)
