+++ src/main/java/org/recordy/server/external/exception/ExternalException.java
@@ -1,0 +1,4 @@
+package org.recordy.server.external.exception;
+
+public class ExternalException {
+}
