+++ src/main/java/org/recordy/server/external/service/S3Service.java
@@ -1,5 +1,6 @@
-package org.recordy.server.external;
+package org.recordy.server.external.service;
 
+import org.recordy.server.external.config.AwsConfig;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
